EDDBK-51 implementation of fetching require scripts on build
