+++ .gitignore
@@ -9,3 +9,8 @@
 !/dist/.gitkeep
 /build/
 ubuntu-xenial-16.04-cloudimg-console.log
+
+npm-debug.log
+
+assets/scripts/*
+!assets/scripts/.gitkeep
